+++ src/main/java/org/morphix/convert/strategy/FinderPathStrategy.java
@@ -72,7 +72,7 @@ public class FinderPathStrategy implements FieldFinderStrategy {
 			} else {
 				resultObject = null;
 			}
-			if (resultField == null) {
+			if (null == resultField) {
 				break;
 			}
 		}