+++ src/test/java/org/morphix/convert/FromIterableTest.java
@@ -53,7 +53,7 @@ class FromIterableTest {
 
 		@Override
 		public int hashCode() {
-			return s == null ? super.hashCode() : s.hashCode();
+			return null == s ? super.hashCode() : s.hashCode();
 		}
 	}
 
@@ -73,7 +73,7 @@ class FromIterableTest {
 
 		@Override
 		public int hashCode() {
-			return s == null ? super.hashCode() : s.hashCode();
+			return null == s ? super.hashCode() : s.hashCode();
 		}
 	}