+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Release Notes
 
+`1.0.32`
+
+- Added `ExecutionWrapper` to `ReschedulingTask`.
+- Added `ExecutionWrappers` utility class which provides basic wrappers.
+- Added `ExecutionWrappers.log` wrapper which logs a start and an end message for the executed task.
+- Added `ExecutionWrappers.time` which times and logs the execution time for the executed task.
+
+---
+
 `1.0.31`
 
 - Changed all `ExecutorService` parameters in `Threads` class to `Executor` for better flexibility.